requirements.txt
binaryornot==0.4.0
chardet==2.3.0
click==6.3
cookiecutter==1.3.0
decorator==4.0.9
Django==1.9.4
django-model-utils==2.4
djangorestframework==3.3.3
future==0.15.2
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
MarkupSafe==0.23
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pyzmq==15.2.0
qtconsole==4.2.0
ruamel.yaml==0.11.7
simplegeneric==0.8.1
terminado==0.6
tornado==4.3
traitlets==4.2.1
whichcraft==0.1.1
